+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
 
+## 1.2.0
+- Added `Config Defaults: Compare Config File...` command to compare current
+	configuration with defaults.
+
 ## 1.1.0
 
 - Files like `settings.default.json` are now treated as JSON with comments to